refactor(backup): extract zip archive writing into backup/archive.rs
